Menggantikan test sehingga bisa ke production

+++ app/tests/conftest.py
@@ -1,5 +1,7 @@
 import asyncio
 
+from app.main import app
+from fastapi.testclient import TestClient
 import pytest
 from google.cloud.firestore import Client
 from google.cloud.storage import Bucket
@@ -9,72 +11,59 @@ from app.database import Base, engine, get_bucket, get_db, get_fs
 from app.schema.user import UserResponse
 from app.services.authentication import create_access_token
 from app.services.user import get_user_by_email
-from app.utils.conftest_utils import delete_all_collection, delete_all_object
+from app.utils.conftest_utils import delete_all_collection, delete_all_object, get_client
 from app.utils.startup import (create_admin_account_if_not_exists,
                                create_test_account_if_not_exists)
-from config import DefaultSettings, get_settings
+from config import DefaultSettings, RunningENV, get_settings
 
 pytest.register_assert_rewrite("app.utils.test")
 
 
 @pytest.fixture(scope="session")
 async def test_db():
-    print("Creating test database")
-    db: Session = next(get_db())
-    fs: Client = next(get_fs())
-    bucket: Bucket = next(get_bucket())
     settings: DefaultSettings = get_settings()
-    Base.metadata.create_all(bind=engine)
-    await create_admin_account_if_not_exists(settings, db)
-    await create_test_account_if_not_exists(settings, db)
-    yield
-    print("Dropping test database")
-    delete_all_object(bucket)
-    delete_all_collection(fs)
-    close_all_sessions()
-    Base.metadata.drop_all(bind=engine)
-
-
-@pytest.fixture(scope="session")
-def admin():
-    db: Session = next(get_db())
-    settings: DefaultSettings = get_settings()
-    user = get_user_by_email(settings.admin_email, db)
-    yield user
-    print("Closing database")
-    close_all_sessions()
-
-
-@pytest.fixture(scope="session")
-def user():
-    db: Session = next(get_db())
-    settings: DefaultSettings = get_settings()
-    user = get_user_by_email(settings.test_account_email, db)
-    yield user
-    print("Closing database")
-    close_all_sessions()
+    if settings.env != RunningENV.TEST_PRODUCTION.value:
+        print("Creating test database")
+        db: Session = next(get_db())
+        fs: Client = next(get_fs())
+        bucket: Bucket = next(get_bucket())
+        Base.metadata.create_all(bind=engine)
+        await create_admin_account_if_not_exists(settings, db)
+        await create_test_account_if_not_exists(settings, db)
+        yield
+        print("Dropping test database")
+        delete_all_object(bucket)
+        delete_all_collection(fs)
+        close_all_sessions()
+        Base.metadata.drop_all(bind=engine)
+    else:
+        yield
 
 
 @pytest.fixture(scope="session")
 def user_token():
-    db: Session = next(get_db())
     settings: DefaultSettings = get_settings()
-    user = get_user_by_email(settings.test_account_email, db)
-    access_token = create_access_token(UserResponse.from_orm(user))
+    client = get_client(settings)
+    response = client.post("/authentications/login", data={
+        "username": settings.test_account_email,
+        "password": settings.test_account_password
+    })
+    resp = response.json()
+    access_token = resp['access_token']
     yield access_token
-    print("Closing database")
-    close_all_sessions()
 
 
 @pytest.fixture(scope="session")
 def admin_token():
-    db: Session = next(get_db())
     settings: DefaultSettings = get_settings()
-    user = get_user_by_email(settings.admin_email, db)
-    access_token = create_access_token(UserResponse.from_orm(user))
+    client = get_client(settings)
+    response = client.post("/authentications/login", data={
+        "username": settings.admin_email,
+        "password": settings.admin_password
+    })
+    resp = response.json()
+    access_token = resp['access_token']
     yield access_token
-    print("Closing database")
-    close_all_sessions()
 
 
 @pytest.fixture(scope='session')
@@ -82,3 +71,10 @@ def event_loop():
     loop = asyncio.get_event_loop_policy().new_event_loop()
     yield loop
     loop.close()
+
+
+@pytest.fixture(scope='session')
+def client():
+    settings = get_settings()
+    client = get_client(settings)
+    yield client
